+++ lib/osm/event_attendance.rb
@@ -1,6 +1,6 @@
 module Osm
 
-  class FlexiRecordData
+  class EventAttendance
     include ::ActiveAttr::MassAssignmentSecurity
     include ::ActiveAttr::Model
 
@@ -32,20 +32,18 @@ module Osm
     def self.from_api(data)
       data.merge!({
         'dob' => data['dob'].nil? ? nil : Osm::parse_date(data['dob'], :ignore_epoch => true),
-        'total' => Osm::to_i_or_nil(data['total'].eql?('') ? nil : data['total']),
-        'completed' => Osm::to_i_or_nil(data['completed'].eql?('') ? nil : data['completed']),
-        'age' => data['age'].eql?('') ? nil : data['age'],
+        'attending' => data['attending'].eql?('Yes'),
       })
 
       new({
         :member_id => Osm::to_i_or_nil(data['scoutid']),
         :grouping_id => Osm::to_i_or_nil(data['patrolid'].eql?('') ? nil : data['patrolid']),
         :fields => data.select { |key, value|
-          ['firstname', 'lastname', 'dob', 'total', 'completed', 'age'].include?(key) || key.to_s.match(/^f_\d+/)
+          ['firstname', 'lastname', 'dob', 'attending'].include?(key) || key.to_s.match(/^f_\d+/)
         }
       })
     end
 
-  end # Class FlexiRecordData
+  end # Class EventAttendance
 
 end # Module