+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Version 0.0.26
 
+  * Event attendance:
+    * Add api.get\_event
+    * Add api.get\_event\_fields
+    * Add api.get\_event\_attendance
   * Fix "uninitialized constant Osm::Api::HTTParty"
 
 ## Version 0.0.25